requirements.txt
 Django
 psycopg2
 djangorestframework
 markdown
 django-filter
 django-taggit